Now a simple irods path is supported as URI.
Just a path like "/zone/home/user/a/b/c.1" is required to instantiate
an IRODSFile. Host, port, user and password are determined via the
account object of the IRODSFileFactoryImpl instance.

All the work in:

	if (uriScheme != null && uriScheme.equals("irods")) {
		...
	}

is for logging. Hence, the if-block could also be commented out.

The instantiation at the end of the method is done with the uri-path,
so it should be sufficient to pass this uri-path in. 

Signed-off-by: Joerg Panzer <panzer@sub.uni-goettingen.de>