StringIndexOutOfBoundsException solved
### I had StringIndexOutOfBoundsException using jargon
```
java.lang.StringIndexOutOfBoundsException: String index out of range: -1
```
when rows are like below
```
[user_name: admin, zone_name: rdss, quota_limit: 1797045, quota_over: -1797045, modify_ts: 01630990563, , user_name: dyjbcswjorsy@test.idr, zone_name: rdss, quota_limit: 100, quota_over: -100, modify_ts: 01630980387]
```
after add "modify_ts" it should stop as you intended, but since it tried to parse blank before next "user_name" it occured Exception as it couldn't find ":" and called substring (0, -1).

It could be solved simply by checking whether idx is a negative number or not. there could be more things to fix if you didn't intended blank space but it would be a simple solution for this exception and won't have any side effects.